"CC BY-SA 3.0 applies to whole" IS NOT QUITE RIGHT
Technically, CC BY-SA 3.0 licence DOES NOT apply to whole unless perhaps I'm willing to sub-license my licence. I'm not willing to do this because of the legal liability implications concerning my liability if I do this. So instead, individual licences remain, and complying with CC BY-SA 3.0 is simply a convenience technique (or shortcut), to complying with the individual licences.